+++ src/serialization.rs
@@ -21,6 +21,10 @@ where
     ordered.serialize(serializer)
 }
 
+pub fn is_true(v: &bool) -> bool {
+    *v
+}
+
 pub fn is_false(v: &bool) -> bool {
     !v
 }